moved connection status strings to align with others
