+++ libs/designer/src/lib/core/templates/utils/helper.ts
@@ -193,31 +193,6 @@ const _sortTemplateManifestEntriesByTitle = (
   }
 };
 
-export const getConnectorResources = (intl: IntlShape) => {
-  return {
-    connected: intl.formatMessage({
-      defaultMessage: 'Connected',
-      id: 'oOGTSo',
-      description: 'Connected text',
-    }),
-    notConnected: intl.formatMessage({
-      defaultMessage: 'Not connected',
-      id: 'WxJJcQ',
-      description: 'Not Connected text',
-    }),
-    authenticated: intl.formatMessage({
-      defaultMessage: 'Authenticated',
-      id: 'iwKxSD',
-      description: 'Connection authenticated text',
-    }),
-    notAuthenticated: intl.formatMessage({
-      defaultMessage: 'Not authenticated',
-      id: 'QdRn5z',
-      description: 'Connection not authenticated text',
-    }),
-  };
-};
-
 export const validateParameterValue = (data: { type: string; value?: string }, required = true): string | undefined => {
   const intl = getIntl();
