+++ libs/designer/src/lib/ui/templates/review/ReviewCreatePanel.tsx
@@ -131,7 +131,6 @@ export const ReviewCreatePanel = () => {
                 connectionKey={connectionKey.replace('_#workflowname#', '')}
                 connectorId={normalizeConnectorId(connections[connectionKey].connectorId ?? '', subscriptionId, location)}
                 hasConnection={mapping[connectionKey] !== undefined}
-                intl={intl}
               />
             ))}
           </div>
